Show modification to scripts/populate.py to accept memory DB URI
The text didn't describe editing the populate.py to accept an optional
``settings`` argument, which _callFUT() in tests.py passed in. This is
used to specify a memory-backed SQLite DB which we can wipe-n-load
during testing.

Show how to replace main() in populate.py per our source code to
accept this argument, and use it to set the DB URI; it also uses our
Page model.