Docs: project.rst: Printf()s can be used for debugging.  Output goes to the server console.
The main point of the second sentence is to setup the reader with
mental context for the 3rd sentence, so that the 3rd sentence sinks
in.  Likewise, the parenthetical in the second sentence about server
startup messages gives the reader some clue as to what the rest of the
sentence it talking about.  I suspect that some readers won't know
what a console is, and the rest will be confused by a server run on a
console.